# Microservice/order-service/src/main/resources/swagger.yaml
openapi: 3.0.3
info:
  title: Order Service API
  version: 1.0.0
  description: API documentation for the Order Service
servers:
  - url: http://localhost:8081
    description: Local development server
paths:
  /orders/placeOrder:
    post:
      summary: Place a new order
      description: Creates a new order and returns the order details.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        "200":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponseDTO'
        "400":
          description: Invalid input

  /orders:
    get:
      summary: Get all orders
      description: Retrieves a list of all placed orders.
      responses:
        "200":
          description: A list of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderResponseDTO'
        "500":
          description: Server error

components:
  schemas:
    Order:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 101
        productId:
          type: integer
          format: int64
          example: 501
        quantity:
          type: integer
          example: 2
        orderDate:
          type: string
          format: date-time
          example: 2025-08-04T10:00:00Z
      required:
        - productId
        - quantity

    ProductDTO:
      type: object
      properties:
        productId:
          type: integer
          format: int64
          example: 501
        name:
          type: string
          example: Wireless Mouse
        price:
          type: number
          format: double
          example: 1499.99

    OrderResponseDTO:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
          example: 101
        product:
          $ref: '#/components/schemas/ProductDTO'
        quantity:
          type: integer
          example: 2
        totalAmount:
          type: number
          format: double
          example: 2999.98
        orderStatus:
          type: string
          example: CONFIRMED